network_monitor/requirements.txt
scapy
python-socketio
requests
